Update pt_br.json (I swear this is the last time)
Now this is the definitive translation for algae thatch, I swear. "Sapé" is not a good translation because in real life it's used to refer to rooftops made with the plant called sapé, so I've changed it to "colmo", wich is a more broad term that can refer to any rooftop made with vegetal materials such as straw or grass.